util/collectives: Track ep per operation not with multicast
The coll_mc (collective multicast) group is assigned an ep
when a new collective operation is initiated.  This will
result in failures of the group is used to start a second
operation on a different endpoint.  The ep should be associated
with the operation, not the multicast group.

Apply a bunch of white space cleanup throughout the collective
code to reduce line lengths to something reasonable.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>