fabtests/inj_complete.c: Make arguments align with the other tests
There is no need for use_recvmsg (-R option) because the msg
flags FI_INJECT and FI_INJECT_COMPLETE are only applicable for
transmit calls, not receive calls.

In normal fabtests, the size is -S and iterations is -I
while in fi_inject_complete FT the size is -m and the
iterations is -O and then -S is used for sendmsg. I added
CS_OPTS and made the test align with the other tests.

send_inj_count has been replaced to opts.iterations.

The tests with msg ep and dgram should be removed to keep
runfabtests simpler since we are setting capabilities to
FI_TAGGED and msg ep and dgram do not support FI_TAGGED.
Due to this, rdm is made the default ep type.

Signed-off-by: Juee Himalbhai Desai <juee.himalbhai.desai@intel.com>