prov/verbs: Format fi_fabric_attr reported through FI_CONNREQ event
Not all fields in the fi_fabric_attr are set when reporting an
fi_info structure as part of an FI_CONNREQ event.  Specifically,
the prov_name and api_version are left unset.  These are usually
set by the core when fi_getinfo is called, but that path is not
taken for CM events.

Copy the values from the passive ep.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>