prov/efa: add state to track FI_HMEM support
Add state to track three things: 1) if libfabric core was able to
initialize the devices we support, 2) whether we support p2p transfer
for hmem, and 3) what the user specified for p2p transfer requirements.

Add checks to validate this state before a transfer.

Finally, add this option to getopt/setopt handlers. Note that p2p
support is still required and enforced through the setopt handler, until
changes to the provider are made to allow copies when possible.

Signed-off-by: Robert Wespetal <wesper@amazon.com>