prov/efa: change FI_HMEM cap reporting behavior around p2p support
Update the provider to report FI_HMEM support only when the required
libraries are present. p2p support will be verified later in the
initialization path where we can verify that a device page can be
registered with the EFA device.

Because the provider only supports FI_HMEM with p2p required, we will
fail the fi_enable call if FI_HMEM is requested and p2p support is
unavailable. This is because we cannot verify whether we can register
GPU pages until the domain is initialized. We also need to allow the
user to change the p2p ep option in the future before enabling the
endpoint.

Remove the gdr sysfs flag check, as it does not accurately report p2p
support in newer kernel driver versions. As mentioned earlier, instead,
allocate and register a CUDA page to determine if it is supported. RDMA
support is still required to be enabled for FI_HMEM as well so leave
those checks in the getinfo path to fail as early as possible.

Signed-off-by: Robert Wespetal <wesper@amazon.com>