contrib/intel: Add test argument to runtests.py
Allow specifying what test should be run by adding an argument.
This will allow the caller to control which test to execute, versus
the current mechanism which tries to run all, with a hidden exclude
option used to filter out tests that shouldn't run.  This change
makes it clearer which test should run.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>