hmem: add support for AWS Trainium devices
Add support for AWS Trainium devices via the AWS Neuron SDK. Writes to
Trainium will be done via the PCIe BAR and reads from Trainium are not
supported at this time.

Despite accessing Trainium memory via the BAR, we still dlopen the
Neuron runtime so that providers allocate a device page and then
register it with the NIC to detect peer to peer support.

Co-authored-by: Venky Natham <vrnatham@amazon.com>
Signed-off-by: Robert Wespetal <wesper@amazon.com>