.github/workflows: temporary workaround for failing ZE build
The Ubuntu distro ZE version was recently updated from 1.4.1 to 1.6.2 which has
an issue in the headers which prevents compilation and causes CI failures.

The issue has been fixed in 1.7.4 but the distro has not been updated yet. This
is a temporary manual workaround to use 1.7.4 instead of the distro version to
avoid CI failures.

Will be removed when the distro gets updated.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>