prov/efa: add AWS Trainium device support
Add support to the EFA provider for AWS Trainium device buffers. RDMA
support and EFA/Trainium peer to peer transactions are required for
Libfabric to support FI_HMEM with Trainium.

Due to the way Libfabric interacts with Trainium buffers, we will force
the LONGREAD protocol for any messages larger than the EFA device MTU
size, so also add a function to check for that.

Co-authored-by: Venky Natham <vrnatham@amazon.com>
Signed-off-by: Robert Wespetal <wesper@amazon.com>