prov/cxi: Work around for MPI disablement of FI_HMEM
When device memory is not used, currently MPI clears
fi_info->caps & FI_HMEM, but not the corresponding tx/rx
attributes caps. OFI utility function do not currently
detect this error, but they should as tx/rx caps must
be a subset of fi_info->caps. Fix this here to coordinate
updating the core and MPI changing how the passed info is
set.

This commit should be reverted when no longer needed.

NETCASSINI-3420

Signed-off-by: Steve Welch <welch@hpe.com>