man/domain: Remove provider requirement to support auto progress
Under certain conditions, such as the shm provider, support for auto
progress would require the allocation of a spinning thread.  The EFA
provider has a similar restriction in that auto progress is not
supported for RMA.  Since existing providers cannot meet this
requirement and never have, remove the text from the man pages.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>