prov/verbs: Store ibv_wr_opcode to use in completion handlnig
The opcode associated with a work request is needed with the
completion in order to properly set the libfabric completion
flags (e.g. FI_RMA, FI_SEND, etc.).  We record the opcode when
posting the request, so that we can use it later to set the
flags.

There are two issues addressed by this patch.  The first is
that verbs does not report the opcode as part of an error
completion.  So, we need to record the opcode and use it if
the request failed.  The second issue is that there was a bug
introduced into libibverbs where the wrong opcode is being
reported.  See issue #7359.  The solution is the same in
both cases.  Record the opcode for the work request and always
use that to set the flags.  The opcode reported in the verbs
completion is ignored.

As part of this change, we rename op_ctx to op_queue to
avoid confusion with the opcode, since that field is only
storing which queue the operation was posted to.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>