prov: Introduce ofi_alter_info_t callback
All implementations of the alter_defaults callback return success.
Remove the need to check for a failure in ofi_prov_check_dup_info
by making it return void, rather than int.  This removes the
assertion that the function did return success.

This adds back the ofi_alter_info_t callback, but in a simpler
form.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>