prov/util: Rename ofi_alter_info_t to ofi_map_info_t
Alter_info is used for 2 purposes.  The first is to map to and
from a util provider's fi_info to a core provider fi_info.  The
second is used to modify the default values stored with an fi_info
structure.  We will separate these functionality to their own
callbacks.  Step 1 is to rename the alter_info callback type to
indicate that the callback performs a mapping between different
providers.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>