prov/verbs: Do not override receive completion opcode
We require verbs to report the correct opcode for receives
in order to distinguish between IBV_WC_RECV and
IBV_WC_RECV_RDMA_WITH_IMM.  There's no way to know which
opcode at the receiver.  So, we hope the bug in verbs that
reports the wrong opcode is limited to send side completions
only.  Use the opcode reported for receive completions.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>