contrib/intel/jenkins: remove dependency on node external access
Currently the CI relies on having all the nodes externally available
so they can each do a clone of libfabric in order to get the Jenkins
code. The code has already been cloned on the head node so we can
just use that.

This patch removes the per-stage checkout and moves the head node
libfabric into a shared directory which the nodes can access.

Signed-off-by: aingerson <alexia.ingerson@intel.com>
Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>