contrib/intel: Add OneCCL tests to CI
Two sets of OneCCL tests, examples and functional, have been
added to the Intel CI.

The test output is too wordy and hence will not be printed
on the console. It is being redirected to /dev/null and the
return code is used to determine if test has passed. OneCCL
team will be adding a new non-verbose option to the tests
for reducing output.

The time taken by a complete CI run is about  23 minutes after
these changes.

Signed-off-by: Juee Himalbhai Desai <juee.himalbhai.desai@intel.com>