prov/shm: use spin lock in smr_region
Previous commit switched to use mutex lock in smr_region,
which has resulted performance degration on various OS.
The degradation is most observable on CentOS 7 and CentOS 8.
For certain workload, the performance can degrade up to 30%.

To addres the issue, this patch switch back to use spinlock
in smr_region.

Signed-off-by: Wei Zhang <wzam@amazon.com>