prov/cxi: Fix operation of fi_alias
Modified operation flags set for an alias EP were not
used as the new default in transmit operations initiated
by it. Fixed so that an alias EP modified flags will
be used with it, and the original with the original
EP/TXC.

This also will provide the groundwork for extending the
software paths that an alias EP will provide.

NETCASSINI-964

Signed-off-by: Steve Welch <welch@hpe.com>