prov/efa: make the RTM packet payload size align on 64 bytes for cuda
When copy data to cuda memory, the performance is better if the
target buffer is alignment on 64 bytes boundary.

Only RTM and DATA packet can carry application data, and DATA packet's
payload is already aligned on 64 bytes.

This patch makes RTM packet's payload size align on 64 bytes when
data is on cuda, and there is more data to send. This ensure each
copy's target buffer starts on 64 bytes boundary, as long as the
applications receiving buffer starts from 64 bytes bounary (which
is usually the case).

Signed-off-by: Wei Zhang <wzam@amazon.com>