prov/cxi: Extend EP alias to support message order
Allow an alias endpoint to override the message order
from the original endpoint. This is an CXI provider
specific mechanism and is initially implemented for
only a normal or alias EP. There are no ordering
guarantees between the original and aliased endpoint.
The message order will not override the message
order for triggered or deferred operations.

This alias behavior is not defined by the libfabric API,
but extends the notion that an alias EP provides is an
alternate software path to the underlying hardware
capabilities. The message order can be modified
using a libfabric control operation on the alias EP.

Example Usage:

    uint64_t msg_order = FI_ORDER_RMA_WAW;

    ret = fi_set_val(&alias_ep->fid, FI_OPT_CXI_SET_MSG_ORDER,
                     &msg_order);
    if (ret)
        goto error;

NETCASSINI-946

Signed-off-by: Steve Welch <welch@hpe.com>