prov/tcp: Avoid accessing partially destroyed endpoints
Re-arrange endpoint cleanup code to ensure that progress functions
(e.g. reading from the cq) cannot access an endpoint that is
in the process of being destroyed.  Remove the EP from all
progress related lists or poll sets prior to closing the socket or
freeing internal data structures.

This avoids a race where we remove a socket from the pollfd set
associated with a CQ, but the endpoint is still accessible from
the CQ ep_list for progressing.  All references from the EQ are
then cleaned up.  However, it's possible that new events could
be added to the EQ, which would later dereference the freed EP.
And since the socket for the EP is still opened, the CQ progress
could conceivably try to process data queued on the socket,
treating such data as a new message, even if it fell in the middle
of a message.

Remove an EP from a CQ or counter atomically, so that the EP is
removed from both the ep list and pollfd set used for progress.
After the EP has been removed from all progress objects, finish
cleanup by removing queued events.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>