configure.ac: rewrite ze checks to minimize ZE dl requirements
Enabling the ZE DL open path only requires having the ZE headers, not the ZE libraries
but the old code was checking for both.

To minimize this requirement, this rewrites the configure code and cleans it up so we
only take one of the two paths - only check for headers with ze_dlopen or check for
headers and the library.

This also fixes an issue where ENABLE_ZE_DLOPEN was only getting defined when enabled
but it is best practice to always set it and check for #if instead of #ifdef

Additionally, the old enable-ze-dlopen check was redundantly checking for libdl when
it was already checked in the beginning as part of the --with-dl check

This also updates the --with-ze help message to better reflect its usage and removes
redundant empty brackets which are not needed.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>