fabtests/runfabtests.py,conftest.py: introduce options.yaml
runfabtests.py is a wrapper of pytest, and the two share
a number of command line options. Currently, to add an
option, the developer will need to modify the parser in
both runfabtests.py and conftest.py, then change runfabtests.py
to pass the options to pytest. This resulted in duplicated
code.

This patch introduced options.yaml file, which is a config
file that contains definition of the shared command line
options. The patch then let runfabtests.py and conftest.py
to load argument from options.yaml.

In the future, one can add an entry to options.yaml to
introduce a new command line option to both runfabtests.py
and libfabric pytest.

Signed-off-by: Wei Zhang <wzam@amazon.com>