configure.ac: rewrite CUDA checks to minimize CUDA dl requirements
Enabling the CUDA DL open path only requires having the CUDA headers, not the libraries
but the old code was checking for both.

To minimize this requirement, this rewrites the configure code and cleans it up so we
only take one of the two paths - only check for headers with cuda_dlopen or check for
headers and the library.

This also fixes an issue where ENABLE_CUDA_DLOPEN was only getting defined when enabled
but it is best practice to always set it and check for #if instead of #ifdef

Additionally, the old enable-cuda-dlopen check was redundantly checking for libdl when
it was already checked in the beginning as part of the --with-dl check

This also updates the --with-cuda help message to better reflect its usage and removes
redundant empty brackets which are not needed.

This changes the psm3 configure to check for have_cuda and not cuda_dlopen since psm3
requires the CUDA runtime and not just the headers.

Signed-off-by: aingerson <alexia.ingerson@intel.com>