configure.ac: cleanup --with-neuron config
Neuron is always built as a dl so needs dlopen but the old check was
redundantly checking for libdl when it was already checked in the
beginning as part of the --with-dl check. This removes that check and
cleans up the neuron check.

This also updates the --with-neuron help message to better reflect its
usage and removes redundant empty brackets which are not needed.

Signed-off-by: aingerson <alexia.ingerson@intel.com>