prov/rxm,verbs: Negotiate credit based flow control support over CM
Credit based flow control only works when both sides support it.
Exchange the information over the connetion manager and disable
flow control if support is missing from either side.

To help this, a new method 'available()' is added to 'ofi_ops_flow_ctrl'
and the 'set_threshold()' method is merged into 'enable()'. In addition,
the hook provider is updated to handle such change.

Backward compatibilty with existing CM protocol is maintained in that
the flow control flag in the CM message occupies a byte previously
used as 0-padding. The new protocol would maintain old semantics if
the flag is 0.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>