prov/efa: add write context packet to rxr_ep->tx_pkt_list
rxr_ep->tx_pkt_list contains a list of TX packets that are waiting for
send completion. Under debug mode, all TX packets in rxr_ep->tx_pkt_list
need to released before destroying the TX packet pool.

The write context packet also need to be released, but it was not added
to tx_pkt_list for shm. This patch fixed the issue.

Signed-off-by: Wei Zhang <wzam@amazon.com>