prov/shm: fix FI_INJECT flag support
Most sendmsg or writemsg calls with FI_INJECT should work
fine without issues, but an assert is added to make sure we don't
use the IOV protocol.

There is an issue when using FI_INJECT and FI_DELIVERY_COMPLETE
or with hmem IPC enabled since those protocols default to an asynchronous
data copy initiated by the remote side. In order to avoid this
incompatbility, add checks in the IPC and IOV paths to make sure that if
we do need FI_DELIVERY_COMPLETE support and immediately copy the data from
the user buffer, use the SAR protocol which supports both flags.

Signed-off-by: aingerson <alexia.ingerson@intel.com>