fabtest/rdm_stress: Add stress test for RDM endpoints
Stress multi-threaded and multi-process access to an RDM endpoint
with abrupt termination conditions.  This will help verify or
expose race conditions in error handling cases.

This patch defines the general framework of operation and
implements most of the client code.  The server side is still
needed.

The general operation is for the client to parse an input file
describing a set of RPCs to perform.  Each RPC involves sending
a short message to the server.  The server will respond by either
sending a message, a tagged message, or issuing an RMA read or
write operation back to the client.  RMA operations are followed
by a response message.

The client side has a parent process and several child processes.
Child processes are forked from the parent, with each child
executing a set of RPC calls to a single server.  The parent
spawns new children as children processes exit.  A child process
may not wait for an RPC response and could exit after sending
a request.

The intent for the server is that the main thread will receive
RPC messages and spawn a thread to handle that request.  The
new thread will generate the RPC response and exit.

The objective is to test a multi-threaded server under load
from multiple clients which are possibly misbehaving.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>