prov/shm: fix 0 byte SAR read
If a 0 byte SAR read is issued, the resp->status will
never be updated since it only gets updated in the
copy_to/from_sar path (which will never be taken if there
is no data to copy). To fix this, update the resp->status
when first progressing the SAR message if the message is
complete.

Signed-off-by: aingerson <alexia.ingerson@intel.com>