prov/cxi: Add SW aligned rdzv to check for minimum length
When issuing a software get for a rendezvous handled by
software (HW could not process or get_en=0); protect
against minimum eager data. The belief is this should not
be required since we force rendezvous eager data to be
larger than cache-line size, but makes the code more
robust.

NETCASSINI-3635

Signed-off-by: Steve Welch <welch@hpe.com>