prov/cxi: Fix for FC during NIC hw-to-sw transition
When the NIC has initiated a hardware to software transition,
properly handle any flow control event (EQ Full, or Request Full)
that may occur before the software transition/on-loading of
unexpected messages is complete. The PtlTE will be re-enabled
in software managed mode when the transition completes and
flow control drops are processed.

NETCASSINI-3648

Signed-off-by: Steve Welch <welch@hpe.com>