prov/efa: refactor rxr_cq_handle_rx_completion()
This patch renamed rxr_cq_handle_rx_completion() to
rxr_cq_complete_rx().

It also moved the sending of ctrl packet, and the
release of rx_entry into this function.

The caller of this function does not need to send
ctrl packet and handle release of rx_entry any more.

Signed-off-by: Wei Zhang <wzam@amazon.com>