prov/linkx: Add FI_LINK capability
Add the FI_LINK capability. This capability is set by providers
which can support linking with other providers.

Linking is a functionality which will be added to later patches.
It allows multiple providers to be linked under the LINKx provider.
These providers can be used in tandem.

The first phase of the work, will allow linking multiple providers,
but only the SHM provider and a second provider will be used.
The SHM provider will be used for on-node transfers, while the other
provider will be used for off-node transfers.

If there are multiple providers available which can be used fore off-node
transfers, then the one with the best performance characteristics
will be used.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>