prov/linkx: add fi_link() API
Add initial implementation of the fi_link() API.
In this patch it handles only one provider. It's behavior
will be exactly like fi_fabric() API.

If there are multiple providers in the passed in list, then
call ofi_create_link() which ends up linking all providers
in the list. Currently this operation is not supported.

Add the initial LINKx implementation.

Although LINKx is a provider, it will not be returned as
part of the matching list in the fi_getinfo() call. It will
behave like it's part of the libfabric infrastructure.
Therefore, it will not implement the <>_getinfo() function. If this
function is called. It'll return EOPNOTSUPP.

Similarly, the fi_fabric() call will not be callable. It will return
EOPNOTSUPP.

Instead the ofi_create_link() will initialize a virtual fabric
and return that to the caller as part of the fi_link() API.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>