core/signal: Restructure fd_signal set/reset
There's a delay between the writing data to the signal fd
socket and that data being available to be read.  This shows
up more frequently on Windows platforms.  We maintain a
state with the signal to indicate if data has been written
to the socket and use that to optimize when to read from the
fd.  However, because of the mentioned race condition, we
aren't guaranteed to read data from the socket even if the
write has already occurred.

This change does a couple of things.  First it replaces the
atomic variables used to track the state with a simpler mutex
and counter.  This just makes the serialization easier to follow.
Second, if we fail to read data from the fd because we hit the
above race condition, we wait until the socket is marked
readable and try again.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>