prov/linkx: Peer table creation
Initialize the LINKx peer table. The peer table uses the OFI
address vector infrastructure.

The peer table will contain an array of peer entries. The size
of the array is determined by the fi_av_attr passed into the fi_av_open()
API: fi_av_attr.count.

Each peer entry will indicate whether the peer is local or remote. It will
also have a list of provider endpoint information:
struct local_prov_entry *eps. This structure contains all the FIDs which are
required for communication.

For on-node peers, it'll have only 1 entry, which would be the SHM provider
endpoint. For remote peers it'll have one or more entries to the providers
which can be used for remote communication.

When the peer is indexed to during traffic operations, then the
appropriate FID for the operation can be dereferenced from the
local_prov_entry field and used in the call to the core provider.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>