prov/linkx: implement av_insert() and av_remove()
Define the LINKx address format.

A LINKx address is an encapsulation of multiple core provider endpoints.
The address format allows specification of multiple providers, and under
each provider multiple addresses.

This will allow LINKx to return a bond of addresses which could in theory
span multiple providers.

When the application calls av_insert() create a peer entry (or update an
existing one). The peer table is referenced by the fid_av.

If the peer is local and we're managing the shm provider, then mark this
peer as only reachable via shared memory, as shared memory will always be
more efficient than all other methods of communication.

If the peer is remote, then we can have multiple addresses provided for
us. The addresses are divided by provider. Go through the address provider
list and insert these addresses into the local endpoint for the same
provider. For example if a node has 2 TCP interfaces, we'll be provided
two TCP addresses, one for each interface. If the local node also has 2
TCP interfaces, they'll be presented as two local provider endpoints, each with
its own Address Vector. In order to be able to reach the peer's interfaces
from either of our local interfaces, we need to insert the remote
addresses into both of the local Address Vector tables for both provider
endpoints.

When the application calls av_remove() remove the peer from the peer table
and all associated core provider addresses from the core providers.

When closing the AV, remove the peer table and clean up all the memory.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>