fabtests/ubertest: fix ubertest writedata completion check
When using SELECTIVE_COMPLETION with the writedata function,
ubertest was assuming that this should not generate a completion.
This is true only if RX_CQ_DATA is set (where an RMA data operation
consumes a recv buffer - if that buffer was set to be suppressed,
that should not generate a completion).
This adds a check to return true for an RMA with data if that mode bit
is not set. If it is set, we should drop into the regular SELECTIVE_COMPLETION
checks to see if the completion will be suppressed or not.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>