prov/efa: allow local read on data received from shm
When receive buffer is on cuda memory and data arrived
in a packet entry, the data is copied to GPU buffer via
local read.

If the packet entry's memory is not registered, a packet
entry need to be allocated from readcopy_pkt_pool, whose
memory is registered, and data will be copied to that entry.

Currently, there is an assertion states that readcopy_pkt_pool
is used only when data is either in unexp_pkt_pool or in ooo_pkt_pool,
bool pools' memory are not registered.

This assertion is wrong because there is a 3rd possiblity that
the data is in shm's RX packet pool, whose memory are also not
registered.

This patch fixed the issue.

Signed-off-by: Wei Zhang <wzam@amazon.com>