fabtest/rdm_stress: Add control file
Add support for control file which contains commands to the server.
Same set of commands is used by all clients.
A sample control file is provided - fabtests/test_configs/ofi_rxm/stress.json.
Details on the format is in the comment section at the top of stress.json.

Add check on number of client connections to 128.

Signed-off-by: Chien Tin Tung <chien.tin.tung@intel.com>