prov/psm3: Add failing tests to exclude file
While running CI for psm3 provider fi_rdm_tagged_peek and
fi_cq_data were failing. Upon manually running these tests
it was found that the tests hang intermittently waiting
for data and once Ctrl-C is pressed to exit, the data is
sent. This looks like a bug in psm3 and an issue has been
opened on github.

Signed-off-by: Juee Himalbhai Desai <juee.himalbhai.desai@intel.com>