prov/linkx: implement send and recv operations
For send/recv operation we currently make an assumption that there are only two
addresses available, a shared memory address if the peer is local, or
another address if the peer is remote.

This will need to be updated to handle Multi-Rail where there could be
multiple addresses we can reach the peer on.

For recv operations, it currently doesn't support handling FI_ADDR_UNSPEC,
as that will require shared receive queues.

Shared receive queues are going to be needed for handling multi-rail, as
you can receive messages on any of the rails.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>