Fix zbcoll setup for integrated testing.
NETCASSINI-3735

The zbcoll and coll simulation code are incompatible, in that zbcoll
requires "special" fi_addrs submitted. This works in the zbcoll test
code, which generates these special addresses, but does not work with
zbcoll-in-coll test code. The special simulation addresses should be
generated internally in zbcoll when requested.

This change primarily modifies _zbcoll_config(). Substantial changes are
required to test ctrl.c to fit the new expectations.

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>